requirements-dev.txt
-r requirements.txt  # This includes the base requirements
ruff
mypy
django-stubs
djangorestframework-stubs
pre-commit
factory-boy
Faker
pytest
pytest-django
